refactor: @EnableJpaAuditing 어노테이션 이동 BackApplication -> AuditorConfig
